refactor(store): split postgres.rs into a postgres/ module directory
